fix: clicking back doesn't re query search (#54)
### TL;DR

Fixes a bug where when you would exit a modal, it would serach for the query below the modal even though nothing changed.

### What changed?
I use prevent default on the button to only make it go browser back, not rerender after clicking.

### Why?
This gives a better user experience and less load on the server.